authentication and tests

+++ osa4/blogs/contollers/blogs.js
@@ -1,14 +1,24 @@
 require('express-async-errors')
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const jwt = require('jsonwebtoken')
+const User = require('../models/user')
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog
+    .find({}).populate('user', { username: 1, name: 1, _id: 1 })
   response.json(blogs)
 })
 
 blogsRouter.post('/', async (request, response) => {
   const body = request.body
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+  if(!decodedToken.id) {
+    return response.status(401).json({ error: 'invalid token' })
+  }
+
+  const user = request.user
 
   if(!request.body.title || !request.body.url) {
     response.status(400).end()
@@ -18,14 +28,29 @@ blogsRouter.post('/', async (request, response) => {
     title : body.title,
     author: body.author,
     url: body.url,
-    likes: body.likes
+    likes: body.likes,
+    user: user._id
   })
 
   const savedBlog = await blog.save()
+  user.blogs = user.blogs.concat(savedBlog._id)
   response.status(201).json(savedBlog)
 })
 
 blogsRouter.delete('/:id', async(request, response) => {
+  const body = request.body
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+  if(!decodedToken.id) {
+    return response.status(401).json({ error: 'invalid token' })
+  }
+
+  const initialBlog = await Blog.findById(request.params.id)
+
+  if(initialBlog.user.toString() !== decodedToken.id.toString()) {
+    return response.status(403).json({ error: 'only creator can delete blog'})
+  }
+
   await Blog.findByIdAndDelete(request.params.id)
 
   response.status(204).end()
